Setup automated testing with GitHub Workflows (#9)
Adding GitHub workflow to run for `main` branch as well as all 
pull-requests to `main`. Example of passing workflow for this
PR: https://github.com/bazel-contrib/unused-jvm-deps/actions/runs/3751014686

__Decision Log__:
  + This was copied from the [rules-template](https://github.com/bazel-contrib/rules-template) repo and modified for our purposes.
  + I've kept the test-matrix code to easily support >1 Bazel versions in the future
  + This is currently pinned to Bazel 5.3.2
  + The Java version is not explicitly specivied

__Note__: I ran into some issues with `bazel test` on Ubuntu 20.10
with Java 19, encountering an annotation in the test that no longer
exists. I've updated the test with a current annotation that
compiles/passes.

Resolves #8